Extract ui components from smtp screen
