+++ portal/src/components/common/ProviderCard.module.css
@@ -1,0 +1,21 @@
+.providerCard {
+  flex: 1;
+  padding: 8px 16px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  border-width: 1px;
+  border-style: solid;
+  cursor: pointer;
+  column-gap: 16px;
+}
+
+.providerCard > * {
+  cursor: inherit;
+}
+
+.columnFull {
+  grid-column: 1 / span 8;
+
+  @apply tablet:col-span-full;
+}
